Faster Hermite generator code
This speeds up the hermf function and reduces the temporary arrays
needed.

Storage is reduced from 3N + 3*N*M to just 3N.

Perforance is about 35% faster.

There are number of techniques used to do this.

Instead of storing the entire P matrix, N × (M+1), we only need to store
two rows, which are used to calculate the next row, which will overwrite
the oldest of the two rows as it's calculated.

Instead of using calculating √dt (2ⁿ n! √π)^-½ from scratch for each k,
as sqrt(dt) / sqrt(sqrt(M_PI) * pow(2, k) * tgamma(k+1)), we can instead
calculate the k=0 value, sqrt(dt / sqrt(M_PI)), and then find the
remaining values by repeatedly dividing by √(2k).

The g vector, e^(-x²/2), can be eliminated by multiplying the initial
rows of P by g.  This them eliminates the need to multiply P by g for
every h calculated.

Instead of calculating Dhₖ using hₖ and hₖ₊₁, use the property of the
derivative of the Hermite polynomial, Pₖ´ = 2k * Pₖ-₁, as part of the
calculation.  While this first appears to be a more complex formula,
most of the terms are already calculated, and it ends up being simple.
This way we avoid the need to calculate the k+1 order of P and h, can
find Dh at the same time as the other values, and avoid the temporary
array Htemp.

Due to the above, it's possible to calculate the P, h, Dh, and Th values
all at once in one pass through an N × M loop, rather than three passes,
which required re-reading the inputs from memory.

This doesn't solve the even N bug, but eliminating the temporary
matrices avoids needing to convert strides when that does get solved.