Use binary search with supplied frequency grid
If fgrid is supplied a linear search is used to find the frequency bin.
This is rather slow.

A test of tfr_spec with len(s)==19670, N = 1024, step = 16, Np = 257,
took ~73 ms.  Adding an fgrid argument about equal to the default fgrid,
np.linspace(0, 0.5, N//2 + 1), increases that time to 505 ms.

After this change, the supplied fgrid version only takes 128 ms.

So this increases the speed almost four times.

Of course it depends on the exact grid and how the points are
distributed.  Most points in the first bin is good for a linear search
while most points in the last is very bad.

The binary search will be more consistent.  E.g., fgrid=np.logspace(-2,
0, N//2+1)/2 is about 738 ms with the linear search but 123 ms with the
binary search.