Remove special case for 2nd order in hermf
The first two orders in hermf were done with explicit formulas rather
than via recurrence, which uses the two previous orders.  I.e., calculating
for k=1 would use data from k=0 and k=-1, and the latter doesn't exist.

But the data from P for k=-1 is immediatly multiplied by zero, so it
doesn't change the result, and there is a valid pointer into P to read
it from (there just isn't real data there).

So the special case code for k=1 can be deleted and main loop can start
at k=1 instead of k=2.

And then the loop to initialize P and be combined with the k=0 loop.