Zero Pkm2 array in hermf
Fixes issue #17

The P values for k = 1..M are computed using the values from the
previous two k.

On the first iteration, k=1, we need the values for k=0, computed before
the k loop started, and for k=-1, which are all zero, but the location
in the P array where they are was never zerod.  So garbage was used, but
that garbage was often all zero by chance, so everything worked.