Check that N is odd for hermf
Remove the N adjustment to even in hermf().  It messes up the array
stride and is totally broken.

We could complicate hermf() to deal with the array stride and number of
points not being the same.  But leaving the last element uninitialized
or setting it to 0 is wrong too, so it doesn't seem worth complicating
the function just to make it easier to get produce bad data without
noticing.

Have the python interface to hermf check that N is odd and raise an
exception if it's not.

Also do this for Np in tfr_spec().

Have mtm_init_herm() check as well, and return NULL in this case.

Add a check to tfr_spec() for mtm_init_herm() returning NULL and raise
an exception.  It still checks for Np not odd so it can raise a better
exception, since the failure cause of mtm_init_herm() returning NULL
isn't known.